requirements.txt
# Core dependencies
streamlit
python-dotenv
numpy
faiss-cpu
Pillow

# Google GenAI API
google-genai==1.0.0

# LangChain versi stabil untuk Python 3.9
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
